requirements.txt
httpx>=0.27
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0
apscheduler>=3.10
loguru>=0.7
scipy>=1.11
fastapi>=0.115
uvicorn>=0.34
# py-clob-client>=0.1  # uncomment when needed in Sprint 6
